DateDiffResult yapılandırıldı,TimeSpanDiffResult ihtiyaç fazlalılığında iptal edildi,ToTimeOnly ve ToPretty foksiyonları eklendi

+++ UD.Core/Helper/Paging/PagingInfo.cs
@@ -3,13 +3,13 @@
     [Serializable]
     public sealed class PagingInfo
     {
-        public long totalcount { get; set; }
-        public long totalpage { get; set; }
+        public int totalcount { get; set; }
+        public int totalpage { get; set; }
         public bool hasnext { get; set; }
         public bool hasprevious { get; set; }
         public PagingInfo() : this(default, default, default, default) { }
-        public PagingInfo(long totalcount, long totalpage, int page) : this(totalcount, totalpage, page < totalpage, page > 1) { }
-        public PagingInfo(long totalcount, long totalpage, bool hasnext, bool hasprevious)
+        public PagingInfo(int totalcount, int totalpage, int page) : this(totalcount, totalpage, page < totalpage, page > 1) { }
+        public PagingInfo(int totalcount, int totalpage, bool hasnext, bool hasprevious)
         {
             this.totalcount = totalcount;
             this.totalpage = totalpage;
